moved currency_to_int func to tools #14
